*: fix many places
- change NetMarble (what the hell is this?) config to 0
- do not encode the password in database
- fix time check when the server start
- fix logout game panic (and then kickout fail)
- do not use EncryptData for packets
- fix ModifyInfo packet encoding on 64bit OS

+++ src/Core/ModifyInfo.h
@@ -202,7 +202,7 @@ typedef struct _SHORTDATA
 typedef struct _LONGDATA
 {
 	BYTE  type;
-	ulong value;
+	DWORD value;
 
 } LONGDATA;
